db type change

+++ database/migrations/2022_04_02_091644_create_purchase_order_items_table.php
@@ -21,7 +21,7 @@ class CreatePurchaseOrderItemsTable extends Migration
             $table->string('item_name');
             $table->unsignedInteger('item_group_id');
             $table->string('item_group_name');
-            $table->unsignedInteger('qty');
+            $table->float('qty', 8, 2);
             $table->unsignedInteger('unit_id');
             $table->float('price', 8, 2);
             $table->string('unit_name');
